Fix compiler warning in assert

+++ src/TriangleMesh-obj.cpp
@@ -232,9 +232,9 @@ bool loadTriangleMeshFromOBJ(TriangleMesh & mesh,
                 assert(indices[vi].vertex_index >= 0);
                 assert(indices[vi].normal_index >= 0);
                 assert(indices[vi].texcoord_index >= 0);
-                assert(indices[vi].vertex_index < meshData.vertices.size());
-                assert(indices[vi].normal_index < meshData.normals.size());
-                assert(indices[vi].texcoord_index < meshData.texcoords.size());
+                assert((ssize_t) indices[vi].vertex_index < meshData.vertices.size());
+                assert((ssize_t) indices[vi].normal_index < meshData.normals.size());
+                assert((ssize_t) indices[vi].texcoord_index < meshData.texcoords.size());
 
                 meshData.indices.vertex.push_back(indices[vi].vertex_index);
                 meshData.indices.normal.push_back(indices[vi].normal_index);
